User-related classes terminated. Code rectification. Relations between tables. Exceptions. Package aspects.

+++ Hello-Mam/src/main/java/com/mindera/HelloMam/services/interfaces/RatingService.java
@@ -10,10 +10,10 @@ import java.util.List;
 
 public interface RatingService {
     List<RatingGetDto> getAllRating();
-    RatingGetDto getRatingById(Integer id) throws RatingNotFoundException;
+    RatingGetDto getRatingById(Long id) throws RatingNotFoundException;
     List<RatingGetDto> getRatingByUserId(Long userId) throws Exception;
-    List <RatingGetDto> getRatingByMediaId(Integer mediaId) throws MediaNotFoundException;
+    List <RatingGetDto> getRatingByMediaId(Long mediaId) throws MediaNotFoundException;
     RatingGetDto addNewRating(RatingCreateDto ratingCreateDto) throws Exception;
-    RatingGetDto updateRating(Integer ratingId, RatingUpdateDto ratingUpdateDto) throws RatingNotFoundException;
+    RatingGetDto updateRating(Long ratingId, RatingUpdateDto ratingUpdateDto) throws RatingNotFoundException;
 
 }
